# perf: unsafe optimizations for verified instruction iteration