--- a/transaction-view/src/bytes.rs
+++ b/transaction-view/src/bytes.rs
@@ -32,6 +32,22 @@ pub fn read_byte(bytes: &[u8], offset: &mut usize) -> Result<u8> {
     value
 }
 
+/// Read a byte and advance the offset without any bounds checks.
+///
+/// * `bytes` - Slice of bytes to read from.
+/// * `offset` - Curernt offset into `bytes`.
+/// * `num_elements` - Number of `T` elements in the slice.
+///
+/// # Safety
+/// 1. `bytes` must be a valid slice of bytes.
+/// 2. `offset` must be a valid offset into `bytes`.
+#[inline(always)]
+pub unsafe fn unchecked_read_byte(bytes: &[u8], offset: &mut usize) -> u8 {
+    let value = *bytes.get_unchecked(*offset);
+    *offset = offset.wrapping_add(1);
+    value
+}
+
 /// Read a compressed u16 from `bytes` starting at `offset`.
 /// If the buffer is too short or the encoding is invalid, return Err.
 /// `offset` is updated to point to the byte after the compressed u16.
@@ -173,6 +189,32 @@ pub unsafe fn read_slice_data<'a, T: Sized>(
     let current_ptr = bytes.as_ptr().wrapping_add(*offset);
     advance_offset_for_array::<T>(bytes, offset, num_elements)?;
     Ok(unsafe { core::slice::from_raw_parts(current_ptr as *const T, usize::from(num_elements)) })
+}
+
+/// Return a reference to the next slice of `T` in the buffer,
+/// and advancing the offset.
+///
+/// * `bytes` - Slice of bytes to read from.
+/// * `offset` - Curernt offset into `bytes`.
+/// * `num_elements` - Number of `T` elements in the slice.
+///
+/// # Safety
+/// 1. `bytes` must be a valid slice of bytes.
+/// 2. `offset` must be a valid offset into `bytes`.
+/// 3. `bytes + offset` must be properly aligned for `T`.
+/// 4. `T` slice must be validly initialized.
+/// 5. The size of `T` is small enough such that a usize will not overflow if
+///    given the maximum slice size (u16::MAX).
+#[inline(always)]
+pub unsafe fn unchecked_read_slice_data<'a, T: Sized>(
+    bytes: &'a [u8],
+    offset: &mut usize,
+    num_elements: u16,
+) -> &'a [T] {
+    let current_ptr = bytes.as_ptr().wrapping_add(*offset);
+    let array_len_bytes = usize::from(num_elements).wrapping_mul(core::mem::size_of::<T>());
+    *offset = offset.wrapping_add(array_len_bytes);
+    unsafe { core::slice::from_raw_parts(current_ptr as *const T, usize::from(num_elements)) }
 }
 
 /// Return a reference to the next `T` in the buffer, checking bounds and