--- a/transaction-view/src/instructions_frame.rs
+++ b/transaction-view/src/instructions_frame.rs
@@ -2,7 +2,7 @@ use {
     crate::{
         bytes::{
             advance_offset_for_array, check_remaining, optimized_read_compressed_u16, read_byte,
-            read_slice_data,
+            unchecked_read_byte, unchecked_read_slice_data,
         },
         result::Result,
     },
@@ -126,7 +126,8 @@ impl<'a> Iterator for InstructionsIterator<'a> {
             // 3. Data ([u8])
 
             // Read the program ID index.
-            let program_id_index = read_byte(self.bytes, &mut self.offset).ok()?;
+            // SAFETY: Offset and length checks have been done in the initial parsing.
+            let program_id_index = unsafe { unchecked_read_byte(self.bytes, &mut self.offset) };
 
             // Move offset to accounts offset - do not re-parse u16.
             self.offset = self.offset.wrapping_add(usize::from(num_accounts_len));
@@ -136,9 +137,10 @@ impl<'a> Iterator for InstructionsIterator<'a> {
             // - The alignment of u8 is 1.
             // - The slice length is checked to be valid.
             // - `u8` cannot be improperly initialized.
-            let accounts =
-                unsafe { read_slice_data::<u8>(self.bytes, &mut self.offset, num_accounts) }
-                    .ok()?;
+            // - Offset and length checks have been done in the initial parsing.
+            let accounts = unsafe {
+                unchecked_read_slice_data::<u8>(self.bytes, &mut self.offset, num_accounts)
+            };
 
             // Move offset to accounts offset - do not re-parse u16.
             self.offset = self.offset.wrapping_add(usize::from(data_len_len));
@@ -148,8 +150,9 @@ impl<'a> Iterator for InstructionsIterator<'a> {
             // - The alignment of u8 is 1.
             // - The slice length is checked to be valid.
             // - `u8` cannot be improperly initialized.
+            // - Offset and length checks have been done in the initial parsing.
             let data =
-                unsafe { read_slice_data::<u8>(self.bytes, &mut self.offset, data_len) }.ok()?;
+                unsafe { unchecked_read_slice_data::<u8>(self.bytes, &mut self.offset, data_len) };
 
             Some(SVMInstruction {
                 program_id_index,