#### Problem
- Instruction iteration can be slow.
- We only ever iterate over data we have already validated has a valid form, yet we rerun all the expensive checks on the data.

#### Summary of Changes
- After initial instruction iteration that does validation, we can skip verification checks and use unsafe optimizations

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->