@apfitzge 2025-03-19
```
receive_and_buffer/transaction_view_max_instructions
                        time:   [433.95 ms 434.33 ms 434.71 ms]
                        thrpt:  [37.689 Kelem/s 37.723 Kelem/s 37.756 Kelem/s]

receive_and_buffer/transaction_view_min_instructions
                        time:   [22.424 ms 22.437 ms 22.450 ms]
                        thrpt:  [729.81 Kelem/s 730.22 Kelem/s 730.64 Kelem/s]

after:

receive_and_buffer/transaction_view_max_instructions
                        time:   [197.68 ms 197.94 ms 198.19 ms]
                        thrpt:  [82.668 Kelem/s 82.774 Kelem/s 82.882 Kelem/s]
                 change:
                        time:   [-54.494% -54.427% -54.356%] (p = 0.00 < 0.05)
                        thrpt:  [+119.09% +119.43% +119.75%]
                        Performance has improved.

receive_and_buffer/transaction_view_min_instructions
                        time:   [20.680 ms 20.710 ms 20.739 ms]
                        thrpt:  [790.00 Kelem/s 791.13 Kelem/s 792.25 Kelem/s]
                 change:
                        time:   [-7.8299% -7.6995% -7.5584%] (p = 0.00 < 0.05)
                        thrpt:  [+8.1764% +8.3417% +8.4950%]
                        Performance has improved.

```