@apfitzge 2025-03-25
rebased with master:
```
receive_and_buffer/transaction_view_max_instructions
                        time:   [174.64 ms 174.89 ms 175.14 ms]
                        thrpt:  [93.551 Kelem/s 93.683 Kelem/s 93.814 Kelem/s]
                 change:
                        time:   [-7.9357% -7.7635% -7.5869%] (p = 0.00 < 0.05)
                        thrpt:  [+8.2097% +8.4170% +8.6197%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) low mild

receive_and_buffer/transaction_view_min_instructions
                        time:   [22.268 ms 22.293 ms 22.318 ms]
                        thrpt:  [734.10 Kelem/s 734.92 Kelem/s 735.75 Kelem/s]
                 change:
                        time:   [-2.2674% -2.0920% -1.9077%] (p = 0.00 < 0.05)
                        thrpt:  [+1.9448% +2.1367% +2.3200%]
                        Performance has improved.
```

It seems some unrelated change has made performance worse in both master and this branch.
I'm guessing one of the recent compiler option changes: #5313.

Indeed, if I additionally revert that commit I get back to 105ktps for the max_instructions case - much closer to the 116ktps I saw when I benched this in #5377.