@KirillLykov 2025-03-26
I guess disabling link-time optimizations lead to this but I don't understand why they have been disabled. If there is serious reason for them to be disabled, I'm not sure if 8% speed up really justifies adding two unsafe functions.