@KirillLykov 2025-03-26
I tried running benchmark using this branch rebased with upstream/master.

Master:
```
receive_and_buffer/transaction_view_max_instructions
    time:   [173.37 ms 174.03 ms 174.95 ms]
    thrpt:  [93.648 Kelem/s 94.147 Kelem/s 94.503 Kelem/s]
```

This PR changes:
```
receive_and_buffer/transaction_view_max_instructions
                        time:   [150.94 ms 151.07 ms 151.21 ms]
                        thrpt:  [108.35 Kelem/s 108.45 Kelem/s 108.55 Kelem/s]
                 change:
                        time:   [-13.655% -13.193% -12.844%] (p = 0.00 < 0.05)
                        thrpt:  [+14.737% +15.198% +15.814%]
                        Performance has improved.
```

So looks actually quite positive.